+++ NEWS.md
@@ -8,6 +8,7 @@ Not released yet.
 
 * New representation: ProMPBehavior
 * New environment: OptimumTrajectoryCurbingObstacles
+* New optimizer: CEMOptimizer
 * Support for Python 3
 
 ### Bugfixes